Refactor /training-planner Simplified the component structure by combining training-studio and training-planner.
