+++ app/login/page.tsx
@@ -8,11 +8,7 @@ export const dynamic = 'force-dynamic'
 export default async function Login() {
   const user = await getUserFromCookie(cookies())
 
-  if (user && !user.admin) {
-    redirect('/training-studio')
-  }
-
-  if (user && user.admin) {
+  if (user) {
     redirect('/training-planner')
   }
