+++ components/calendar/CalendarMedium.tsx
@@ -2,11 +2,18 @@ import React from 'react'
 import {useLockBodyScroll} from '@/hooks'
 import {useIsMobile} from '@/hooks'
 
-export function CalendarMedium({children}: {children: React.ReactNode}) {
+export function CalendarMedium({
+  children,
+  isAdmin,
+}: {
+  children: React.ReactNode
+  isAdmin?: boolean
+}) {
   const isMobile = useIsMobile()
   useLockBodyScroll()
 
-  if (isMobile) {
+  // TODO: Admin view is not optimised for mobile
+  if (isMobile && !isAdmin) {
     return null
   }
