+++ middleware.ts
@@ -2,11 +2,9 @@ import {NextRequest, NextResponse} from 'next/server'
 import {jwtVerify} from 'jose'
 import {RequestCookies} from 'next/dist/compiled/@edge-runtime/cookies'
 
-// Unauthenticated users are redirected to /login from all (training-app) routes in app/(training-app)/layout.tsx
-// Middleware logic intercepts requests before the above redirect
+// Unauthenticated users are redirected to /login from all (training-app) group routes in app/(training-app)/layout.tsx
 
-const adminRoutes = ['/clients', '/register', '/training-planner']
-const clientRoutes = ['/training-studio']
+const adminRoutes = ['/clients', '/register']
 
 export default async function middleware(req: NextRequest) {
   const {pathname} = req.nextUrl
@@ -14,8 +12,8 @@ export default async function middleware(req: NextRequest) {
   if (adminRoutes.some(route => pathname.startsWith(route))) {
     const JWTPayload = await getJWTPayload(req.cookies)
 
-    // This avoids redirecting unauthenticated users first to /training-studio
-    // then to /login by app/(training-app)/layout.tsx
+    // This avoids redirecting unauthenticated users first to /training-planner
+    // then to /login via app/(training-app)/layout.tsx
     if (!JWTPayload) {
       req.nextUrl.pathname = '/login'
 
@@ -23,24 +21,6 @@ export default async function middleware(req: NextRequest) {
     }
 
     if (!JWTPayload.admin) {
-      req.nextUrl.pathname = '/training-studio'
-
-      return NextResponse.redirect(req.nextUrl)
-    }
-  }
-
-  if (clientRoutes.some(route => pathname.startsWith(route))) {
-    const JWTPayload = await getJWTPayload(req.cookies)
-
-    // This avoids redirecting unauthenticated users first to /training-planner
-    // then to /login by app/(training-app)/layout.tsx
-    if (!JWTPayload) {
-      req.nextUrl.pathname = '/login'
-
-      return NextResponse.redirect(req.nextUrl)
-    }
-
-    if (JWTPayload.admin) {
       req.nextUrl.pathname = '/training-planner'
 
       return NextResponse.redirect(req.nextUrl)
